refactor: 💡 Move out `createA11yElement` function
